requirements-prod.txt
# TriadRank ATS - Production Requirements
# 3-Tier Resume Scoring and Ranking System

# Core ML Framework
torch>=2.0.0

# Transformers and NLP
transformers>=4.30.0
spacy>=3.7.2,<3.8.0

# PDF Processing (for real-world resume extraction)
pypdf>=3.9.0
pdfplumber>=0.10.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.23.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Utilities
pyyaml>=6.0
python-multipart>=0.0.6
pydantic>=2.0.0
